move session creation logic to delivery session service
