+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/ActionPlanService.kt
@@ -92,12 +92,7 @@ class ActionPlanService(
   fun approveActionPlan(id: UUID, user: AuthUser): ActionPlan {
     val actionPlan = getActionPlan(id)
 
-    actionPlan.referral.approvedActionPlan?. let {
-      val sessions = deliverySessionRepository.findAllByActionPlanId(id)
-      deliverySessionService.createUnscheduledSessionsForActionPlan(actionPlan, sessions.count())
-    } ?: run {
-      deliverySessionService.createUnscheduledSessionsForActionPlan(actionPlan)
-    }
+    deliverySessionService.createUnscheduledSessionsForActionPlan(actionPlan)
 
     actionPlan.approvedAt = OffsetDateTime.now()
     actionPlan.approvedBy = authUserRepository.save(user)
