+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/ActionPlanServiceTest.kt
@@ -244,29 +244,7 @@ internal class ActionPlanServiceTest {
     assertThat(approvedActionPlan.approvedAt).isNotNull
     assertThat(approvedActionPlan.approvedBy).isEqualTo(authUser)
     verify(actionPlanEventPublisher).actionPlanApprovedEvent(same(actionPlan))
-    verify(deliverySessionService).createUnscheduledSessionsForActionPlan(same(actionPlan), same(0))
-  }
-
-  @Test
-  fun `approve when there is a previously approved action plan passess previous sessions`() {
-    val newActionPlanId = UUID.randomUUID()
-    val referral = referralFactory.createSent()
-    val previouslyApprovedActionPlan = actionPlanFactory.createApproved(numberOfSessions = 2, referral = referral)
-    val newActionPlan = actionPlanFactory.createSubmitted(id = newActionPlanId, numberOfSessions = 2, referral = referral)
-    referral.actionPlans = mutableListOf(previouslyApprovedActionPlan, newActionPlan)
-
-    val authUser = AuthUser("CRN123", "auth", "user")
-    whenever(actionPlanRepository.findById(newActionPlanId)).thenReturn(of(newActionPlan))
-    whenever(authUserRepository.save(any())).then(AdditionalAnswers.returnsFirstArg<AuthUser>())
-    whenever(actionPlanRepository.save(any())).then(AdditionalAnswers.returnsFirstArg<ActionPlan>())
-    whenever(deliverySessionRepository.findAllByActionPlanId(any())).thenReturn(listOf(deliverySessionFactory.createAttended(), deliverySessionFactory.createAttended()))
-
-    val approvedActionPlan = actionPlanService.approveActionPlan(newActionPlanId, authUser)
-    assertThat(approvedActionPlan.approvedAt).isNotNull
-    assertThat(approvedActionPlan.approvedBy).isEqualTo(authUser)
-    verify(actionPlanEventPublisher).actionPlanApprovedEvent(same(newActionPlan))
-    verify(deliverySessionService).createUnscheduledSessionsForActionPlan(same(newActionPlan), same(2))
-    verify(deliverySessionRepository, times(1)).findAllByActionPlanId(any())
+    verify(deliverySessionService).createUnscheduledSessionsForActionPlan(same(actionPlan))
   }
 
   @Test
