+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/DeliverySessionsServiceTest.kt
@@ -31,6 +31,7 @@ import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.repository.Del
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.util.ActionPlanFactory
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.util.AuthUserFactory
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.util.DeliverySessionFactory
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.util.ReferralFactory
 import java.time.OffsetDateTime
 import java.util.UUID
 import javax.persistence.EntityExistsException
@@ -48,6 +49,7 @@ internal class DeliverySessionsServiceTest {
   private val actionPlanFactory = ActionPlanFactory()
   private val deliverySessionFactory = DeliverySessionFactory()
   private val authUserFactory = AuthUserFactory()
+  private val referralFactory = ReferralFactory()
 
   private val deliverySessionsService = DeliverySessionService(
     deliverySessionRepository, actionPlanRepository,
@@ -71,14 +73,20 @@ internal class DeliverySessionsServiceTest {
   }
 
   @Test
-  fun `create unscheduled sessions where some sessions are already created`() {
-    val actionPlan = actionPlanFactory.create(numberOfSessions = 3)
-    whenever(deliverySessionRepository.findAllByActionPlanIdAndSessionNumber(eq(actionPlan.id), any())).thenReturn(null)
-    whenever(authUserRepository.save(actionPlan.createdBy)).thenReturn(actionPlan.createdBy)
+  fun `create unscheduled sessions where there is a previously approved action plan`() {
+    val newActionPlanId = UUID.randomUUID()
+    val referral = referralFactory.createSent()
+    val previouslyApprovedActionPlan = actionPlanFactory.createApproved(numberOfSessions = 2, referral = referral)
+    val newActionPlan = actionPlanFactory.createSubmitted(id = newActionPlanId, numberOfSessions = 3, referral = referral)
+    referral.actionPlans = mutableListOf(previouslyApprovedActionPlan, newActionPlan)
+
+    whenever(deliverySessionRepository.findAllByActionPlanId(any())).thenReturn(listOf(deliverySessionFactory.createAttended(), deliverySessionFactory.createAttended()))
+    whenever(deliverySessionRepository.findAllByActionPlanIdAndSessionNumber(eq(newActionPlan.id), any())).thenReturn(null)
+    whenever(authUserRepository.save(newActionPlan.createdBy)).thenReturn(newActionPlan.createdBy)
     whenever(deliverySessionRepository.save(any())).thenAnswer { it.arguments[0] }
 
-    deliverySessionsService.createUnscheduledSessionsForActionPlan(actionPlan, 1)
-    verify(deliverySessionRepository, times(2)).save(any())
+    deliverySessionsService.createUnscheduledSessionsForActionPlan(newActionPlan)
+    verify(deliverySessionRepository, times(1)).save(any())
   }
 
   @Test
